linear temperature update. softmax layer bug fix, vis filter script. data_transformer test code bug fix. NCM dist modified.

+++ src/caffe/data_transformer.cpp
@@ -508,7 +508,8 @@ void DataTransformer<Dtype>::Transform(const cv::Mat& cv_img,
 		  Dtype* transformed_data = transformed_blob->mutable_cpu_data()+transformed_blob->offset(ncrop);
 		  int top_index;
 		  for (int h = 0; h < height; ++h) {
-			  const uchar* ptr = cv_cropped_img.ptr<uchar>(h);
+			  //const uchar* ptr = cv_cropped_img.ptr<uchar>(h);
+			  const float* ptr = cv_cropped_img.ptr<float>(h);
 			  int img_index = 0;
 			  for (int w = 0; w < width; ++w) {
 				  for (int c = 0; c < img_channels; ++c) {
@@ -615,7 +616,8 @@ void DataTransformer<Dtype>::Transform(const cv::Mat& cv_img,
 		  Dtype* transformed_data = transformed_blob->mutable_cpu_data() + transformed_blob->offset(nrotate);
 		  int top_index;
 		  for (int h = 0; h < height; ++h) {
-			  const uchar* ptr = cv_cropped_img.ptr<uchar>(h);
+			  //const uchar* ptr = cv_cropped_img.ptr<uchar>(h);
+			  const float* ptr = cv_cropped_img.ptr<float>(h);
 			  int img_index = 0;
 			  for (int w = 0; w < width; ++w) {
 				  for (int c = 0; c < img_channels; ++c) {
